feat(dapp-client): add color mode
