+++ .gitignore
@@ -68,5 +68,6 @@ storage.json
 audit.json
 gl-dependency-scanning-report.json
 
-src/alert/alert-templates.ts
-src/alert/wallet-lists.ts
+src/ui/alert/alert-templates.ts
+src/ui/toast/toast-templates.ts
+src/ui/alert/wallet-lists.ts
