+++ assets/alert/alert-container.html
@@ -1,4 +1,4 @@
-<div id="beacon-alert-modal-{{id}}" class="beacon-alert__base animated fadeIn theme__dark">
+<div id="beacon-alert-modal-{{id}}" class="beacon-alert__base animated fadeIn">
   <div class="beacon-modal__wrapper">
     <div class="beacon-modal__header">
       <svg
