+++ assets/toast/toast.css
@@ -1,0 +1,178 @@
+:root {
+  --animation-duration: 300ms;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes fadeOut {
+  from {
+    opacity: 1;
+  }
+  to {
+    opacity: 0;
+  }
+}
+
+.animated {
+  animation-duration: var(--animation-duration);
+  animation-fill-mode: both;
+}
+
+.fadeIn {
+  animation-name: fadeIn;
+}
+
+.fadeOut {
+  animation-name: fadeOut;
+}
+
+.beacon-toast__base {
+  position: absolute;
+  top: 16px;
+  right: 16px;
+  z-index: 2147483000;
+  margin: 0 auto;
+  border-radius: 8px;
+  overflow: hidden;
+  box-shadow: 0 12px 24px 0 rgba(0, 0, 0, 0.1);
+  flex-direction: column;
+}
+
+.theme__light.beacon-toast__base {
+  background: #fff;
+}
+
+.theme__dark.beacon-toast__base {
+  background: #27334c;
+}
+
+.theme__dark p {
+  color: #6183ff;
+}
+
+a {
+  text-decoration: none;
+  color: #3880ff;
+}
+
+a svg {
+  width: 12px;
+  margin-left: 2px;
+}
+
+a:visited {
+  color: #3880ff;
+}
+
+hr {
+  height: 1px;
+  color: rgba(0, 0, 0, 0.12);
+  background-color: rgba(0, 0, 0, 0.12);
+  border: none;
+  width: 100%;
+  margin: 0;
+}
+
+.beacon-toast__content {
+  justify-content: space-between;
+}
+
+.beacon-toast__content p,
+.beacon-toast__action__item {
+  align-items: center;
+}
+
+.beacon-toast__content,
+.beacon-toast__base,
+.beacon-toast__action__item,
+.beacon-toast__content p {
+  display: flex;
+}
+
+.beacon-toast__content,
+.beacon-toast__action__item {
+  padding: 0 16px;
+  font-family: Roboto, Helvetica, sans-serif;
+}
+
+.beacon-toast__action__item {
+  font-size: 14px;
+}
+
+.beacon-toast__action__item p {
+  margin-right: 8px;
+  align-items: center;
+}
+
+.beacon-toast__action__item__subtitle {
+  min-width: 88px;
+  color: rgba(0, 0, 0, 0.54);
+}
+
+.beacon-toast__content__img {
+  width: 24px;
+  height: 24px;
+  margin: 0 4px 0 6px;
+}
+
+.beacon-toast__more {
+  width: 14px;
+  padding-left: 32px;
+}
+
+.beacon-toast__more--action {
+  margin: 8px 0;
+}
+
+.progress-line,
+.progress-line:before {
+  height: 3px;
+  width: 100%;
+  margin: 0;
+}
+.progress-line {
+  background-color: #a7c4f7;
+  display: -webkit-flex;
+  display: flex;
+}
+.progress-line:before {
+  background-color: #3880ff;
+  content: '';
+  -webkit-animation: running-progress 2s cubic-bezier(0.4, 0, 0.2, 1) infinite;
+  animation: running-progress 2s cubic-bezier(0.4, 0, 0.2, 1) infinite;
+}
+@-webkit-keyframes running-progress {
+  0% {
+    margin-left: 0px;
+    margin-right: 100%;
+  }
+  50% {
+    margin-left: 25%;
+    margin-right: 0%;
+  }
+  100% {
+    margin-left: 100%;
+    margin-right: 0;
+  }
+}
+@keyframes running-progress {
+  0% {
+    margin-left: 0px;
+    margin-right: 100%;
+  }
+  50% {
+    margin-left: 25%;
+    margin-right: 0%;
+  }
+  100% {
+    margin-left: 100%;
+    margin-right: 0;
+  }
+}
