+++ example-dapp.html
@@ -56,6 +56,17 @@
     <br /><br />
     ---
     <br /><br />
+    <span>
+      Color Mode:
+      <span id="activeColorMode"></span>
+    </span>
+    <br /><br />
+    <button id="color-mode-light">Set color mode "light"</button>
+    <br /><br />
+    <button id="color-mode-dark">Set color mode "dark"</button>
+    <br /><br />
+    ---
+    <br /><br />
     <button id="contractCall">Contract Call</button>
     <br /><br />
     ---
@@ -83,6 +94,14 @@
         })
       }
       updateActiveAccount()
+
+      // Display the active account in the UI
+      const updateColorMode = () => {
+        client.getColorMode().then((colorMode) => {
+          document.getElementById('activeColorMode').innerHTML = colorMode
+        })
+      }
+      updateColorMode()
 
       // Initiate a delegate operation
       const delegate = () => {
@@ -271,6 +290,18 @@
       document.getElementById('2requests').addEventListener('click', () => {
         setTimeout(sendToSelf, 0)
         setTimeout(sendToSelf, 1000)
+      })
+
+      // Add event listener to the button
+      document.getElementById('color-mode-light').addEventListener('click', () => {
+        client.setColorMode('light')
+        updateColorMode()
+      })
+
+      // Add event listener to the button
+      document.getElementById('color-mode-dark').addEventListener('click', () => {
+        client.setColorMode('dark')
+        updateColorMode()
       })
     </script>
   </body>
