+++ scripts/generate-wallet-list.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
 import * as fs from 'fs'
 import { NetworkType } from '../src/types/beacon/NetworkType'
-import { ExtensionApp, WebApp, App, AppBase, DesktopApp } from '../src/alert/Pairing'
+import { ExtensionApp, WebApp, App, AppBase, DesktopApp } from '../src/ui/alert/Pairing'
 
 export const extensionList: ExtensionApp[] = [
   {
@@ -117,7 +117,7 @@ const convert = <T extends AppBase>(list: T[]): Promise<T[]> => {
 }
 
 const createLists = async () => {
-  const ALERT_DEST_DIR = path.join(PKG_DIR, 'src', 'alert')
+  const ALERT_DEST_DIR = path.join(PKG_DIR, 'src', 'ui', 'alert')
 
   const extensionListWithInlinedLogo = await convert(extensionList)
   const desktopListWithInlinedLogo = await convert(desktopList)
@@ -159,7 +159,7 @@ const createLists = async () => {
 
 const createAlert = async () => {
   const ALERT_SRC_DIR = path.join(PKG_DIR, 'assets', 'alert')
-  const ALERT_DEST_DIR = path.join(PKG_DIR, 'src', 'alert')
+  const ALERT_DEST_DIR = path.join(PKG_DIR, 'src', 'ui', 'alert')
 
   const css = (await readFile(path.join(ALERT_SRC_DIR, 'alert.css'))).toString('utf-8')
   const pairCss = (await readFile(path.join(ALERT_SRC_DIR, 'alert-pair.css'))).toString('utf-8')
@@ -189,5 +189,27 @@ const createAlert = async () => {
   )
 }
 
+const createToast = async () => {
+  const TOAST_SRC_DIR = path.join(PKG_DIR, 'assets', 'toast')
+  const TOAST_DEST_DIR = path.join(PKG_DIR, 'src', 'ui', 'toast')
+
+  const css = (await readFile(path.join(TOAST_SRC_DIR, 'toast.css'))).toString('utf-8')
+
+  let html = (await readFile(path.join(TOAST_SRC_DIR, 'toast.html'))).toString('utf-8')
+
+  const x = {
+    default: {
+      html: html,
+      css: css
+    }
+  }
+
+  writeFile(
+    path.join(TOAST_DEST_DIR, 'toast-templates.ts'),
+    `export const toastTemplates = ${JSON.stringify(x)}`
+  )
+}
+
 createLists()
 createAlert()
+createToast()
