+++ src/colorMode.ts
@@ -1,0 +1,9 @@
+import { ColorMode } from './types/ColorMode'
+
+let colorMode: ColorMode = ColorMode.LIGHT
+
+export const setColorMode = (mode: ColorMode): void => {
+  colorMode = mode
+}
+
+export const getColorMode = (): ColorMode => colorMode
