+++ src/types/ColorMode.ts
@@ -1,0 +1,4 @@
+export enum ColorMode {
+  LIGHT = 'light',
+  DARK = 'dark'
+}
